database migration with checks

+++ database/migrations/2016_09_21_224211_create_game_relationship_for_each_player.php
@@ -12,21 +12,23 @@ class CreateGameRelationshipForEachPlayer extends Migration
      */
     public function up()
     {
-        $tournaments = \App\Models\Championship\Relation\PlayerRelation::where('relation_type', '=', \App\Models\Championship\Tournament::class)->get();
-        foreach ($tournaments as $key => $tournament){
-            $the_tournament = \App\Models\Championship\Tournament::where('id', '=', $tournament->relation_id)->first();
-            if($the_tournament == null){
-                $tournament->delete();
-            }else {
-                $p_id = $tournament->player_id;
-                $g_id = $the_tournament->game_id;
-                $relation = new \App\Models\Championship\Relation\PlayerRelation();
-                $relation->player_id = $p_id;
-                $relation->relation_id = $g_id;
-                $relation->relation_type = \App\Models\Championship\Game::class;
-                $exists = \App\Models\Championship\Game::find($g_id)->hasPlayerID($p_id);
-                if (!$exists) {
-                    $relation->save();
+        if (Schema::connection('mysql_champ')->hasTable('Tournament') and Schema::connection('mysql_champ')->hasTable('player_relations') ) {
+            $tournaments = \App\Models\Championship\Relation\PlayerRelation::where('relation_type', '=', \App\Models\Championship\Tournament::class)->get();
+            foreach ($tournaments as $key => $tournament) {
+                $the_tournament = \App\Models\Championship\Tournament::where('id', '=', $tournament->relation_id)->first();
+                if ($the_tournament == null) {
+                    $tournament->delete();
+                } else {
+                    $p_id = $tournament->player_id;
+                    $g_id = $the_tournament->game_id;
+                    $relation = new \App\Models\Championship\Relation\PlayerRelation();
+                    $relation->player_id = $p_id;
+                    $relation->relation_id = $g_id;
+                    $relation->relation_type = \App\Models\Championship\Game::class;
+                    $exists = \App\Models\Championship\Game::find($g_id)->hasPlayerID($p_id);
+                    if (!$exists) {
+                        $relation->save();
+                    }
                 }
             }
         }
